Initial prototype
This app listens for GitHub "status" updates which are published when a
PR's CI status is updated. When this occurs, it extracts the build ID
from the status and queries the Travis API for job status. If a job
defines a CHECK_NAME env variable, that job will be surfaced as a check
and its status will be reported. Additionally, once a job completes if
its log output contains a fenced `output` block, the contents of the
block is assumed to be an [`output`
object](https://developer.github.com/v3/checks/runs/#output-object), and
will be attached to the check payload.