Remove streaming logic and await the full response
The streaming logic does not appear to be reliable, even when
considering all of the edge cases and implementing a deadline timer to
ensure that the promise is always eventually resolved. The main
advantage of the approach was that we could avoid storing the entire
(relatively large) log file in memory and just scan through it for the
lines we cared about, but after numerous attempts to make it completely
reliable it seems like we'll just keep running into edge cases that add
to the complexity.

The refactored approach involves storing the full response in memory as
a string and splitting it to parse the response line by line. We use an
iterator-based version of `split()` to avoid having to allocate
additional space for the resulting array and copying all of the log
lines into it.